DeepfakeBench/training/wma/requirements.txt
grpcio>=1.59.0
grpcio-tools>=1.59.0
protobuf>=4.24.0
asyncio>=3.4.3
pathlib>=1.0.1